it is what it is

+++ Panel-Tester/src/main/kotlin/dev/slimevr/testing/actions/ExecuteCommandAction.kt
@@ -2,6 +2,10 @@ package dev.slimevr.testing.actions
 
 import dev.slimevr.testing.TestResult
 import dev.slimevr.testing.TestStatus
+import java.io.IOException
+import java.util.concurrent.CompletableFuture
+import java.util.logging.Logger
+import java.util.regex.Pattern
 
 /**
  * Executes console command and tests output for patterns.
@@ -11,14 +15,16 @@ import dev.slimevr.testing.TestStatus
  */
 class ExecuteCommandAction(
     testName: String,
-    successPatterns: Array<String>,
-    failurePatterns: Array<String>,
+    successPatterns: Array<Pattern>,
+    failurePatterns: Array<Pattern>,
     command: String,
     private val timeout: Long
 ) : MatchingAction(testName, successPatterns, failurePatterns) {
 
+    private val logger: Logger = Logger.getLogger("ExecuteCommandAction")
+
     private val processBuilder: ProcessBuilder =
-        ProcessBuilder(command.split(" ")).redirectError(ProcessBuilder.Redirect.PIPE).redirectOutput(
+        ProcessBuilder(command.split(" ")).redirectErrorStream(true).redirectOutput(
             ProcessBuilder.Redirect.PIPE
         )
 
@@ -27,22 +33,24 @@ class ExecuteCommandAction(
         if (log.isNotBlank())
             fullLog.add(log)
         var logStart = fullLog.size
+        logger.info("Starting process: " + processBuilder.command().joinToString(" "))
         val process = processBuilder.start()
         val endTime = System.currentTimeMillis() + timeout
-        val errorReader = process.errorReader()
         val inputReader = process.inputReader()
         var testResult = TestStatus.TESTING
         var matchedString = ""
-        while (process.isAlive) {
-            when (val line = errorReader.readLine()) {
-                null -> break
-                else -> fullLog.add(line)
-            }
-            when (val line = inputReader.readLine()) {
-                null -> break
-                else -> fullLog.add(line)
+        var destroyed = false
+        do {
+            try {
+                when (val line = inputReader.readLine()) {
+                    null -> break
+                    else -> if (line.isNotBlank()) fullLog.add(line)
+                }
+            } catch(ex: IOException) {
+                ex.message?.let { fullLog.add(it) }
             }
             fullLog.subList(logStart, fullLog.size).forEach {
+                logger.info(it)
                 val result = matchString(it)
                 if ((result == MatchResult.SUCCESS) && (testResult != TestStatus.ERROR)) {
                     matchedString = it
@@ -51,17 +59,18 @@ class ExecuteCommandAction(
                     testResult = TestStatus.ERROR
                 }
             }
-            if (timeout > 0 && endTime < System.currentTimeMillis()) {
+            if (!destroyed && timeout > 0 && endTime < System.currentTimeMillis()) {
                 process.destroyForcibly()
                 fullLog.add("Forcibly destroying process after ${timeout / 1000}s")
                 testResult = TestStatus.ERROR
-                break
+                destroyed = true
             }
             logStart = fullLog.size
-        }
+        } while(process.isAlive)
+        //fullLog.add("Process ended with ${process.exitValue()} exit code")
         // If we didn't pass the test yet, it's a failure
         if (testResult == TestStatus.TESTING) {
-            testResult = if (successPatterns.isEmpty() && failurePatterns.isEmpty() && process.exitValue() == 0)
+            testResult = if (successPatterns.isEmpty() && failurePatterns.isEmpty())// && process.exitValue() == 0)
                 TestStatus.PASS
             else
                 TestStatus.ERROR
@@ -72,6 +81,6 @@ class ExecuteCommandAction(
             startTime,
             System.currentTimeMillis(),
             matchedString,
-            fullLog.joinToString { "\n" })
+            fullLog.joinToString("\n"))
     }
 }
